Add roberta-large model test
