+++ tests/jax/models/roberta/tester.py
@@ -5,26 +5,32 @@
 from typing import Dict, Sequence
 
 import jax
-import pytest
 from flax import linen as nn
-from infra import ModelTester, RunMode
+from infra import ComparisonConfig, ModelTester, RunMode
 from transformers import AutoTokenizer, FlaxRobertaForMaskedLM
-
-MODEL_PATH = "FacebookAI/roberta-base"
-
-# ----- Tester -----
 
 
 class FlaxRobertaForMaskedLMTester(ModelTester):
     """Tester for Roberta model on a masked language modeling task."""
 
+    # TODO(mrakita): Add tests for other variants.
+
+    def __init__(
+        self,
+        model_name: str,
+        comparison_config: ComparisonConfig = ComparisonConfig(),
+        run_mode: RunMode = RunMode.INFERENCE,
+    ) -> None:
+        self._model_name = model_name
+        super().__init__(comparison_config, run_mode)
+
     # @override
     def _get_model(self) -> nn.Module:
-        return FlaxRobertaForMaskedLM.from_pretrained(MODEL_PATH)
+        return FlaxRobertaForMaskedLM.from_pretrained(self._model_name)
 
     # @override
     def _get_input_activations(self) -> Sequence[jax.Array]:
-        tokenizer = AutoTokenizer.from_pretrained(MODEL_PATH)
+        tokenizer = AutoTokenizer.from_pretrained(self._model_name)
         inputs = tokenizer("Hello <mask>.", return_tensors="np")
         return inputs["input_ids"]
 
@@ -39,33 +45,3 @@ class FlaxRobertaForMaskedLMTester(ModelTester):
     # @ override
     def _get_static_argnames(self):
         return ["train"]
-
-
-# ----- Fixtures -----
-
-
-@pytest.fixture
-def inference_tester() -> FlaxRobertaForMaskedLMTester:
-    return FlaxRobertaForMaskedLMTester()
-
-
-@pytest.fixture
-def training_tester() -> FlaxRobertaForMaskedLMTester:
-    return FlaxRobertaForMaskedLMTester(RunMode.TRAINING)
-
-
-# ----- Tests -----
-
-
-@pytest.mark.xfail(reason="failed to legalize operation 'stablehlo.reduce_window'")
-def test_roberta_inference(
-    inference_tester: FlaxRobertaForMaskedLMTester,
-):
-    inference_tester.test()
-
-
-@pytest.mark.skip(reason="Support for training not implemented")
-def test_flax_roberta_training(
-    training_tester: FlaxRobertaForMaskedLMTester,
-):
-    training_tester.test()
